requirements.txt
asttokens==2.0.5
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
et-xmlfile==1.1.0
executing==0.8.2
icecream==2.1.1
idna==3.3
mypy-extensions==0.4.3
numpy==1.22.1
openpyxl==3.0.9
pandas==1.4.0
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.4.1
Pygments==2.11.2
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
tomli==1.2.3
typing-extensions==4.0.1
urllib3==1.26.8
